+++ action.sh
@@ -6,10 +6,11 @@ echo "[Zygisk - Frida Gadget] redeploy start"
 echo
 
 . "$MODDIR/deploy_gadget.sh"
+DEPLOY_LOG_STDOUT=0
 deploy_gadget "$MODDIR"
 
 echo
-echo "[Zygisk - Frida Gadget] deploy.log"
+echo "[Zygisk - Frida Gadget] deploy result"
 echo "----------------------------------------"
 
 if [ -f "$MODDIR/deploy.log" ]; then